changed column and started view and controller for game index

+++ db/migrate/20170215200857_create_games.rb
@@ -6,7 +6,7 @@ class CreateGames < ActiveRecord::Migration[5.0]
       t.integer :max_players
       t.integer :playing_time
       t.integer :year_published
-      t.integer :bgg_rating
+      t.decimal :bgg_rating
       t.string :image
       t.string :thumbnail
       t.timestamps
